refactor: Update Body1 component to use HTML5 semantic tag
