+++ src/components/typography/body/body2.tsx
@@ -1,7 +1,7 @@
 import { ReactNode } from "react";
-import { typography } from "../../../typography/typography";
+import { typography } from "../../../typography";
 
 function Body2({ children }: { children: ReactNode }) {
-  return <body style={typography.body.body2}>{children}</body>;
+  return <p style={typography.body.body2}>{children}</p>;
 }
 export default Body2;
