+++ src/components/typography/body/body4.tsx
@@ -1,7 +1,7 @@
 import { ReactNode } from "react";
-import { typography } from "../../../typography/typography";
+import { typography } from "../../../typography";
 
 function Body4({ children }: { children: ReactNode }) {
-  return <body style={typography.body.body4}>{children}</body>;
+  return <p style={typography.body.body4}>{children}</p>;
 }
 export default Body4;
